Remove HostScanner and HostFinger to their own files
