+++ monkey/infection_monkey/network/HostScanner.py
@@ -1,0 +1,7 @@
+from abc import ABCMeta, abstractmethod
+
+
+class HostScanner(object, metaclass=ABCMeta):
+    @abstractmethod
+    def is_host_alive(self, host):
+        raise NotImplementedError()
