+++ monkey/infection_monkey/network/tcp_scanner.py
@@ -2,7 +2,9 @@ from itertools import zip_longest
 from random import shuffle
 
 import infection_monkey.config
-from infection_monkey.network import HostScanner, HostFinger
+import infection_monkey.network
+import infection_monkey.network.HostFinger
+import infection_monkey.network.HostScanner
 from infection_monkey.network.tools import check_tcp_ports, tcp_port_to_service
 
 __author__ = 'itamar'
@@ -10,8 +12,7 @@ __author__ = 'itamar'
 BANNER_READ = 1024
 
 
-class TcpScanner(HostScanner, HostFinger):
-
+class TcpScanner(infection_monkey.network.HostScanner.HostScanner, infection_monkey.network.HostFinger.HostFinger):
     _SCANNED_SERVICE = 'unknown(TCP)'
 
     def __init__(self):
